feat: adding gate for authorizing user before using API

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,13 +8,14 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
 
     use CanLoadRelationships;
 
-    private $relations = ['event', 'user'];
+    private array $relations = ['event', 'user'];
 
     public function index(Event $event)
     {
@@ -56,6 +57,8 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
+        Gate::authorize('delete-attendee', [$event, $attendee]);
+
         $attendee->delete();
 
         return response()->json([
